% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.gamma.covmat.R
\name{AD.gamma.covmat}
\alias{AD.gamma.covmat}
\title{AD.gamma.covmat}
\usage{
AD.gamma.covmat(n, shape)
}
\arguments{
\item{n}{number of eigenvalues}

\item{shape}{shape parameter for Gamma distribution}
}
\value{

}
\description{
AD.gamma.covmat
}
