% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.gamma.eigen.R
\name{AD.gamma.eigen}
\alias{AD.gamma.eigen}
\title{AD.gamma.eigen}
\usage{
AD.gamma.eigen(n, shape)
}
\arguments{
\item{n}{number of eigenvalues}

\item{shape}{shape parameter for Gamma distribution}
}
\value{

}
\description{
AD.gamma.eigen
}
