% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.R
\name{AD.normal}
\alias{AD.normal}
\alias{AD.gamma}
\alias{AD.logistic}
\alias{AD.laplace}
\alias{AD.weibull}
\alias{AD.exp}
\title{EDF Statistic A^2 for a Given Distribution}
\usage{
AD.normal(x, parameter = estimate.normal(x))

AD.gamma(x, parameter = estimate.gamma(x))

AD.logistic(x, parameter = estimate.logistic(x))

AD.laplace(x, parameter = estimate.laplace(x))

AD.weibull(x, parameter = estimate.weibull(x))

AD.exp(x, parameter = estimate.exp(x))
}
\arguments{
\item{x}{A random sample.}

\item{parameter}{Parameters of a given distribution.}
}
\value{
Anderson-Darling statistic of a uniform sample.
}
\description{
Compute the Anderson-Darling goodness-of-fit statistic A^2 for an iid sample,
x, to test for the given distribution with parameters unknown.
Estimate parameters by ML using \code{EDFtest} mle function by default.
}
\details{

}
\examples{
x1=rnorm(n=100,mean=0,sd=1)
CvM.normal(x1)
CvM.normal(x1,c(0,1))

x2=rgamma(n=100,shape=1,scale=1)
CvM.gamma(x2)
CvM.gamma(x2,parameter=c(1,1))

x3=rlogis(n=100,location=0,scale=1)
CvM.logistic(x3)
CvM.logistic(x3,c(0,1))

x4= rmutil::rlaplace(n=100,m=0,s=1)
CvM.laplace(x4)
CvM.laplace(x4,c(0,1))

x5=rweibull(n=100,shape=1,scale=1)
CvM.weibull(x5)
CvM.weibull(x5,c(1,1))

x6=rexp(n=100,rate=1/2)
CvM.exp(x6)
CvM.exp(x6,parameter=2)

}
