% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.regression.statistic.R
\name{AD.regression}
\alias{AD.regression}
\alias{AD.extremevalue.regression}
\title{Anderson-Darling statistic for regression}
\usage{
AD.extremevalue.regression(
  y,
  x,
  parameter = estimate.extremevalue.regression(y, x)
)
}
\arguments{
\item{y}{response variable}

\item{x}{explanatory variables}

\item{parameter}{estimates of regression parameters}
}
\value{

}
\description{

}
\examples{

}
\seealso{

}
