% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.pvalue.R
\name{AD.uniform.pvalue}
\alias{AD.uniform.pvalue}
\alias{AD.normal.pvalue}
\alias{AD.gamma.pvalue}
\alias{AD.logistic.pvalue}
\alias{AD.laplace.pvalue}
\alias{AD.weibull.pvalue}
\alias{AD.exp.pvalue}
\title{P-value of Anderson-Darling statistic}
\usage{
AD.uniform.pvalue(a, neig = 100, verbose = FALSE)

AD.normal.pvalue(a, neig = 100, verbose = FALSE)

AD.gamma.pvalue(a, shape, neig = 100, verbose = FALSE)

AD.logistic.pvalue(a, neig = 100, verbose = FALSE)

AD.laplace.pvalue(a, neig = 100, verbose = FALSE)

AD.weibull.pvalue(a, neig = 100, verbose = FALSE)

AD.exp.pvalue(a, neig = 100, verbose = FALSE)
}
\arguments{
\item{a}{Anderson-Darling statistic \eqn{A^2} with a given distribution.}

\item{neig}{Number of eigenvalues used for \code{\link[CompQuadForm]{imhof}}.}

\item{verbose}{Logical; if `TRUE`, print warning messages.}

\item{shape}{The shape parameter of Gamma distribution.}
}
\value{
P-value of the Anderson-Darling statistic.
}
\description{
Compute the P-value of the given Anderson-Darling statistic \eqn{A^2}
using \code{\link[CompQuadForm]{imhof}} function in \code{CompQuadForm}.
}
\examples{
x0=runif(n=100,min=-1,max=1)
asq0 = AD.uniform(x0)
AD.uniform.pvalue(asq0)

x1=rnorm(n=100,mean=0,sd=1)
asq1 = AD.normal(x1)
AD.normal.pvalue(asq1)

x2=rgamma(n=100,shape=1,scale=1)
asq2 = AD.gamma(x2)
AD.gamma.pvalue(asq2,1)

x3=rlogis(n=100,location=0,scale=1)
asq3 = AD.logistic(x3)
AD.logistic.pvalue(asq3)

x4=rmutil::rlaplace(n=100,m=0,s=1)
asq4 = AD.laplace(x4)
AD.laplace.pvalue(asq4)

x5=rweibull(n=100,shape=1,scale=1)
asq5 = AD.weibull(x5)
AD.weibull.pvalue(asq5)

x6=rexp(n=100,rate=1/2)
asq6 = AD.exp(x6)
AD.exp.pvalue(asq6)
}
