% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.weibull.R
\name{AD.weibull}
\alias{AD.weibull}
\title{EDF statistics A^2 for Weibull Distribution}
\usage{
AD.weibull(x, parameter = estimate.weibull(x))
}
\arguments{
\item{x}{random sample}

\item{parameter}{parameter of Weibull distribution}
}
\value{
AD.weibull gives Anderson-Darling statistic of a uniform sample.
}
\description{
Compute Anderson-Darling statistic A^2 for an iid sample, x, to test for the Weibull distribution with parameters unknown.
Estimate parameters by ML using "estimate.weibull" by default.
}
\examples{
x= rweibull(1000,1)
AD.weibull(x)
AD.weibull(x,c(1,1))
}
