% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.exp.R
\name{cdf.exp}
\alias{cdf.exp}
\title{Probability Integral Transforms for Exponential Distribution}
\usage{
cdf.exp(x, theta)
}
\arguments{
\item{x}{random sample}

\item{theta}{parameter of Exponential distribution}
}
\value{
cdf.exp gives probability integral transforms for dataset following Weibull distribution
}
\description{
Compute the probability integral transforms of the random sample x in a Exponential distribution
}
\examples{
x= rexp(100,1)
cdf.exp(x,1)
}
