% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{cdf.gamma.regression}
\alias{cdf.gamma.regression}
\title{Probability Integral Transforms for Gamma Regression}
\usage{
cdf.gamma.regression(y, x, thetahat, link = "log")
}
\arguments{
\item{y}{response variable}

\item{x}{explanatory variables}

\item{theta.hat}{estimates of coefficients}
}
\value{
cdf.gamma.regression gives probability integral transforms for y in a gamma log-linear regression model
}
\description{
Computes the probability integral transforms of the responses y
in a gamma log-linear regression model: log(E(y)) = x beta
}
