% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.regression.statistic.R
\name{cdf.normal.regression}
\alias{cdf.normal.regression}
\title{Computes the probability integral transforms of the responses y
in a normal linear regression model}
\usage{
cdf.normal.regression(y, x, theta.hat)
}
\description{
Computes the probability integral transforms of the responses y
in a normal linear regression model
}
