% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.laplace.R
\name{CvM.laplace}
\alias{CvM.laplace}
\title{EDF statistics W^2 for Laplace Distribution}
\usage{
CvM.laplace(x, parameter = estimate.laplace(x))
}
\arguments{
\item{x}{random sample}

\item{parameter}{parameter of Laplace distribution}
}
\value{
CvM.laplace gives Cramer-von Mises statistic of a uniform sample.
}
\description{
Compute Cramer-von Mises statistic W^2 for an iid sample, x, to test for the Laplace distribution with parameters unknown.
Estimate parameters by ML using "estimate.laplace" by default.
}
\examples{
library(rmutil)
x= rlaplace(1000,0,1)
CvM.laplace(x)
CvM.laplace(x,c(0,1))
}
