% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.normal.R
\name{CvM.normal}
\alias{CvM.normal}
\title{EDF statistics W^2 for Normal Distribution}
\usage{
CvM.normal(x, parameter = estimate.normal(x))
}
\arguments{
\item{x}{random sample}

\item{parameter}{parameter of Normal distribution}
}
\value{
CvM.normal gives Cramer-von Mises statistic of a uniform sample.
}
\description{
Compute Cramer-von Mises statistic W^2 for an iid sample, x, to test for the Normal distribution with parameters unknown.
Estimate parameters by ML using "estimate.normal" by default.
}
\examples{
x= rnorm(1000)
CvM.normal(x)
CvM.normal(x,c(0,1))
}
