% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.regression.statistic.R
\name{CvM.regression}
\alias{CvM.regression}
\alias{CvM.normal.regression}
\alias{CvM.gamma.regression}
\alias{CvM.logistic.regression}
\alias{CvM.laplace.regression}
\alias{CvM.weibull.regression}
\alias{CvM.extremevalue.regression}
\alias{CvM.exp.regression}
\title{Cramer-von Mises statistic for regression}
\usage{
CvM.normal.regression(y, x, fit.intercept = TRUE)

CvM.gamma.regression(y, x, fit, fit.intercept = TRUE, link = "log")

CvM.logistic.regression(y, x, fit.intercept = TRUE)

CvM.laplace.regression(y, x, fit.intercept = TRUE)

CvM.weibull.regression(y, x, fit.intercept = TRUE)

CvM.extremevalue.regression(y, x, fit.intercept = TRUE)

CvM.exp.regression(y, x, fit, fit.intercept = TRUE, link = "log")
}
\arguments{
\item{y}{response variable}

\item{x}{explanatory variables}

\item{parameter}{estimates of regression parameters}
}
\value{

}
\description{

}
\examples{
n = 500
p = 3
beta = c(1,2,3)
x = rnorm(n*(p))
#x = c(rep(1,n),x)
x = matrix(x,n,p)
# OLS regression
mean = x\%*\%(beta)
#apply the link to get the mean
#generate data with that mean,

sd = 2
y =rnorm(n,mean=mean,sd=sd)
estimate.normal.regression(x=x,y=y,fit.intercept=TRUE)
CvM.normal.regression(x=x,y=y,fit.intercept=TRUE) # mean(cvm)=.06


}
\seealso{

}
