% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.pvalue.R
\name{CvM.uniform.pvalue}
\alias{CvM.uniform.pvalue}
\alias{CvM.normal.pvalue}
\alias{CvM.gamma.pvalue}
\alias{CvM.logistic.pvalue}
\alias{CvM.laplace.pvalue}
\alias{CvM.weibull.pvalue}
\alias{CvM.exp.pvalue}
\title{P-value of Cramer-von Mises statistic}
\usage{
CvM.uniform.pvalue(w, neig = 100, verbose = FALSE)

CvM.normal.pvalue(w, neig = 100, verbose = FALSE)

CvM.gamma.pvalue(w, shape, neig = 100, verbose = FALSE)

CvM.logistic.pvalue(w, neig = 100, verbose = FALSE)

CvM.laplace.pvalue(w, neig = 100, verbose = FALSE)

CvM.weibull.pvalue(w, neig = 100, verbose = FALSE)

CvM.exp.pvalue(w, neig = 100, verbose = FALSE)
}
\arguments{
\item{w}{Cramér-von Mises statistic \eqn{W^2} with a given distribution.}

\item{neig}{Number of eigenvalues used for \code{\link[CompQuadForm]{imhof}}.}

\item{verbose}{Logical; if TRUE, print warning messages.}

\item{shape}{The shape parameter of Gamma distribution.}
}
\value{
P-value of the given Cramér-von Mises statistic.
}
\description{
Compute the P-value of the given Cramér-von Mises statistic \eqn{W^2}
using \code{\link[CompQuadForm]{imhof}} function in \code{CompQuadForm}.
}
\examples{
x0=runif(n=100,min=-1,max=1)
wsq0 = CvM.uniform(x0)
CvM.uniform.pvalue(wsq0)

x1=rnorm(n=100,mean=0,sd=1)
wsq1 = CvM.normal(x1)
CvM.normal.pvalue(wsq1)

x2=rgamma(n=100,shape=1,scale=1)
wsq2 = CvM.gamma(x2)
CvM.gamma.pvalue(wsq2,1)

x3=rlogis(n=100,location=0,scale=1)
wsq3 = CvM.logistic(x3)
CvM.logistic.pvalue(wsq3)

x4= rmutil::rlaplace(n=100,m=0,s=1)
wsq4 = CvM.laplace(x4)
CvM.laplace.pvalue(wsq4)

x5=rweibull(n=100,shape=1,scale=1)
wsq5 = CvM.weibull(x5)
CvM.weibull.pvalue(wsq5)

x6=rexp(n=100,rate=1/2)
wsq6 = CvM.exp(x6)
CvM.exp.pvalue(wsq6)
}
