% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.statistic.R
\name{CvM.uniform}
\alias{CvM.uniform}
\alias{CvM.normal}
\alias{CvM.gamma}
\alias{CvM.logistic}
\alias{CvM.laplace}
\alias{CvM.weibull}
\alias{CvM.exp}
\title{EDF Statistic W^2 for a Given Distribution}
\usage{
CvM.uniform(x, parameter = estimate.uniform(x))

CvM.normal(x, parameter = estimate.normal(x))

CvM.gamma(x, parameter = estimate.gamma(x))

CvM.logistic(x, parameter = estimate.logistic(x))

CvM.laplace(x, parameter = estimate.laplace(x))

CvM.weibull(x, parameter = estimate.weibull(x))

CvM.exp(x, parameter = estimate.exp(x))
}
\arguments{
\item{x}{A random sample.}

\item{parameter}{Parameters of a given distribution.}
}
\value{
Cramer-von Mises statistic of a uniform sample.
}
\description{
Compute Cramer-von Mises goodness-of-fit statistic W^2 for an iid sample, x, to test for
the given distribution with parameters unknown. Estimate parameters by ML
using \code{EDFtest} mle function by default.
}
\details{

}
\examples{
x1=rnorm(n=100,mean=0,sd=1)
CvM.normal(x1)
CvM.normal(x1,parameter=c(0,1))

x2=rgamma(n=100,shape=1,scale=1)
CvM.gamma(x2)
CvM.gamma(x2,parameter=c(1,1))

x3=rlogis(n=100,location=0,scale=1)
CvM.logistic(x3)
CvM.logistic(x3,parameter=c(0,1))

x4= rmutil::rlaplace(n=100,m=0,s=1)
CvM.laplace(x4)
CvM.laplace(x4,parameter=c(0,1))

x5=rweibull(n=100,shape=1,scale=1)
CvM.weibull(x5)
CvM.weibull(x5,parameter=c(1,1))

x6=rexp(n=100,rate=1/2)
CvM.exp(x6)
CvM.exp(x6,parameter=2)
}
