% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.weibull.covmat.R
\name{CvM.weibull.covmat}
\alias{CvM.weibull.covmat}
\title{CvM.weibull.covmat}
\usage{
CvM.weibull.covmat(n)
}
\arguments{
\item{n}{number of eigenvalues}
}
\value{

}
\description{
CvM.weibull.covmat
}
