% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.normal.R
\name{estimate.normal}
\alias{estimate.normal}
\title{MLE for Normal Distribution}
\usage{
estimate.normal(x)
}
\arguments{
\item{x}{random sample}
}
\value{
estimated location and scale parameters of the Normal distribution.
}
\description{
Estimate location and scale parameters of the Normal distribution by the method of maximum likelihood.
}
\examples{
x=rnorm(10)
estimate.normal(x)
}
