% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other.Regression.R
\name{fisher.information.normal}
\alias{fisher.information.normal}
\title{Fisher Information for Normal Regression Model}
\usage{
fisher.information.normal(x)
}
\arguments{
\item{x}{matrix of covariates, normally x will contain an intercept term, that is a column of 1s}
}
\description{
The value returned is for sigma=1.  In general the FI must be divided by sigma^2.
}
