% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.exp.bootstrap.R
\name{gof.exp.bootstrap}
\alias{gof.exp.bootstrap}
\title{EDF Goodness-of-Fit tests for Exponential Distribution by Bootstrap}
\usage{
gof.exp.bootstrap(x, M = 10000)
}
\arguments{
\item{x}{random sample}

\item{M}{number of bootstrap}
}
\value{
gof.exp.bootstrap computes Cramer-von Mises and Anderson-Darling statistics and their P-values by bootstrap.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Exponential
parameters, compute probability integral transforms, and computes Cramer-von Mises
and Anderson-Darling statistics. P-values are calculated by M bootstrap.
}
\examples{
x=rexp(1000,1/2)
gof.exp.bootstrap(x,M=1000)
gof.exp(x)
}
