% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.exp.R
\name{gof.exp}
\alias{gof.exp}
\title{EDF Goodness-of-Fit tests for Exponential Distribution}
\usage{
gof.exp(x, print = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{random sample}

\item{print}{Logical, if TRUE print statistics and P-values}

\item{verbose}{Logical, if TRUE print every step}
}
\value{
gof.exp computes Cramer-von Mises and Anderson-Darling statistics and their P-values.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Exponential
parameters, compute probability integral transforms, and computes Cramer-von Mises
and Anderson-Darling statistics and their P-values
}
\examples{
x = rexp(100,1/2)
gof.exp(x)
gof.exp(x,print=TRUE,verbose=TRUE)
}
