% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.gamma.R
\name{gof.gamma}
\alias{gof.gamma}
\title{EDF Goodness-of-Fit tests for Gamma Distribution}
\usage{
gof.gamma(x, print = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{random sample}

\item{print}{Logical, if TRUE print statistics and P-values}

\item{verbose}{Logical, if TRUE print every step}
}
\value{
gof.gamma computes Cramer-von Mises and Anderson-Darling statistics and their P-values.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Gamma
parameters, compute probability integral transforms, and computes Cramer-von Mises
and Anderson-Darling statistics and their P-values
}
\examples{
x = rgamma(100,1)
gof.gamma(x)
gof.gamma(x,print=TRUE,verbose=TRUE)
}
