% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof.normal}
\alias{gof.normal}
\alias{gof.normal.bootstrap}
\alias{gof.gamma}
\alias{gof.gamma.bootstrap}
\alias{gof.logistic}
\alias{gof.logistic.bootstrap}
\alias{gof.laplace}
\alias{gof.laplace.bootstrap}
\alias{gof.weibull}
\alias{gof.weibull.bootstrap}
\alias{gof.exp}
\alias{gof.exp.bootstrap}
\alias{gof.uniform.bootstrap}
\title{EDF Goodness-of-Fit Tests for a Given Distribution}
\usage{
gof.normal(x, print = TRUE, verbose = FALSE)

gof.normal.bootstrap(x, M = 10000)

gof.gamma(x, print = TRUE, verbose = FALSE)

gof.gamma.bootstrap(x, M = 10000)

gof.logistic(x, print = TRUE, verbose = FALSE)

gof.logistic.bootstrap(x, M = 10000)

gof.laplace(x, print = TRUE, verbose = FALSE)

gof.laplace.bootstrap(x, M = 10000)

gof.weibull(x, print = TRUE, verbose = FALSE)

gof.weibull.bootstrap(x, M = 10000)

gof.exp(x, print = TRUE, verbose = FALSE)

gof.exp.bootstrap(x, M = 10000)

gof.uniform.bootstrap(x, stat = y, m = 10000)
}
\arguments{
\item{x}{A random sample.}

\item{print}{Logical; if TRUE print both statistics and P-values.}

\item{verbose}{verbose Logical; if TRUE, print warning messages.}

\item{M}{Number of bootstrap, 10000 by default.}

\item{stat}{}
}
\value{
Cramer-von Mises, Anderson-Darling and Watson's statistics and their P-values.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Normal
parameters, compute probability integral transforms, and computes Cramer-von Mises,
Anderson-Darling and Watson's statistics and their P-values using either `imhof`
method or bootstrap.
}
\details{

}
\examples{
x1=rnorm(n=1000,mean=0,sd=1)
gof.normal(x1)
gof.normal.bootstrap(x1,M=1000)

x2=rgamma(n=1000,shape=1,scale=1)
gof.gamma(x)
gof.gamma.bootstrap(x2,M=1000)

x3=rlogis(n=1000,location=0,scale=1)
gof.logistic(x3)
gof.logistic.bootstrap(x3,M=1000)

x4= rmutil::rlaplace(n=1000,m=0,s=1)
gof.laplace(x4)
gof.laplace.bootstrap(x4,M=1000)

x5 = rweibull(100,1)
gof.weibull(x5)
gof.weibull(x5,print=TRUE,verbose=TRUE)

x6=rweibull(1000,1)
gof.weibull(x6)
gof.weibull.bootstrap(x6,M=1000)

x7 = rexp(100,1/2)
gof.exp(x7)
gof.exp(x7,print=TRUE,verbose=TRUE)

x8=rexp(1000,1/2)
gof.exp.bootstrap(x8,M=1000)
gof.exp(x8)
}
