% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.statistics.only.R
\name{gof.statistics.only}
\alias{gof.statistics.only}
\title{EDF Goodness-of-Fit tests for Statistics Only}
\usage{
gof.statistics.only(pit, AD = TRUE, CvM = TRUE, Watson = TRUE)
}
\arguments{
\item{pit}{A vector of probability integral transforms.}

\item{a2}{Logical; if TRUE print Anderson-Darling statistic}

\item{w2}{Logical; if TRUE print Cramer-von Mises statistic}

\item{u2}{Logical; if TRUE print Watson statistic}
}
\value{
`gof.statistics.only` gives Anderson-Darling, Cramer-von Mises or Watson statistics
}
\description{
This function takes in a vector of probability integral transforms
and computes whichever of the three statistics is asked for.
By default A^2,W^2, and U^2 are computed
}
\details{

}
\examples{
x= runif(100)
gof.statistics.only(x)
}
