% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof.uniform}
\alias{gof.uniform}
\alias{gof.uniform.bootstrap}
\alias{gof.normal}
\alias{gof.normal.bootstrap}
\alias{gof.gamma}
\alias{gof.gamma.bootstrap}
\alias{gof.logistic}
\alias{gof.logistic.bootstrap}
\alias{gof.laplace}
\alias{gof.laplace.bootstrap}
\alias{gof.weibull}
\alias{gof.weibull.bootstrap}
\alias{gof.exp}
\alias{gof.exp.bootstrap}
\title{Goodness-of-Fit tests based on EDF}
\usage{
gof.uniform(x, print = TRUE, verbose = FALSE)

gof.uniform.bootstrap(x, M = 10000)

gof.normal(x, print = TRUE, verbose = FALSE)

gof.normal.bootstrap(x, M = 10000)

gof.gamma(x, print = TRUE, verbose = FALSE)

gof.gamma.bootstrap(x, M = 10000)

gof.logistic(x, print = TRUE, verbose = FALSE)

gof.logistic.bootstrap(x, M = 10000)

gof.laplace(x, print = TRUE, verbose = FALSE)

gof.laplace.bootstrap(x, M = 10000)

gof.weibull(x, print = TRUE, verbose = FALSE)

gof.weibull.bootstrap(x, M = 10000)

gof.exp(x, print = TRUE, verbose = FALSE)

gof.exp.bootstrap(x, M = 10000)
}
\arguments{
\item{x}{A random sample.}

\item{print}{Logical; if TRUE print both statistics and P-values.}

\item{verbose}{verbose Logical; if TRUE, print warning messages.}

\item{M}{Number of bootstrap, 10000 by default.}
}
\value{
Cramer-von Mises, Anderson-Darling and Watson's statistics and their P-values.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Normal
parameters, compute probability integral transforms, and computes Cramer-von Mises,
Anderson-Darling and Watson's statistics and their P-values using either `imhof`
method or bootstrap.
}
\examples{
x0=runif(n=100,min=-1,max=1)
gof.uniform(x0)
gof.uniform.bootstrap(x0,M=100)

x1=rnorm(n=100,mean=0,sd=1)
gof.normal(x1)
gof.normal.bootstrap(x1,M=100)

x2=rgamma(n=100,shape=1,scale=1)
gof.gamma(x2)
gof.gamma.bootstrap(x2,M=100)

x3=rlogis(n=100,location=0,scale=1)
gof.logistic(x3)
gof.logistic.bootstrap(x3,M=100)

x4= rmutil::rlaplace(n=100,m=0,s=1)
gof.laplace(x4)
gof.laplace.bootstrap(x4,M=100)

x5=rweibull(n=100,shape=1,scale=1)
gof.weibull(x5)
gof.weibull.bootstrap(x5,M=100)

x6=rexp(n=100,rate=1/2)
gof.exp(x6)
gof.exp.bootstrap(x6,M=100)
}
