% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.normal.R
\name{score.normal}
\alias{score.normal}
\title{Score Function for Normal Distribution}
\usage{
score.normal(x, theta)
}
\arguments{
\item{x}{random sample}

\item{theta}{parameter of Normal distribution}
}
\value{
score.normal returns an n by 2 matrix whose ith row refers to the ith data point.
}
\description{
Compute the score function for the two parameter Normal distribution
}
