% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.normal.regression.R
\name{score.normal.regression}
\alias{score.normal.regression}
\title{Score Function for Simple Linear Regression Model}
\usage{
score.normal.regression(y, x, theta.hat)
}
\arguments{
\item{y}{the univariate response}

\item{x}{the design matrix}

\item{theta.hat}{MLEs of parameters}
}
\value{
Column 1 of the output is the sigma component of the score. Columns 2 to p+1
are the components of the score corresponding to the regression coefficients.
}
\description{
computes an n by p+1 matrix containing the components of the score function
in a linear regression of y on the n x p covariate matrix x.
The ith row of the output corresponds to the ith data point.
}
\examples{
m = matrix(rnorm(20),10,2)
X = cbind(rep(1,10),m)
y = rnorm(10)
fit = lm(y~X-1)
sigma = sqrt(mean(fit$residual^2))
theta.hat =  c(sigma,fit$coefficients)
score.normal.regression(y,X,theta.hat=theta.hat)
round(apply(.Last.value,2,sum))
}
