% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.exp.R
\name{Watson.exp}
\alias{Watson.exp}
\title{EDF statistics U^2 for Exponential Distribution}
\usage{
Watson.exp(x, parameter = estimate.exp(x))
}
\arguments{
\item{x}{random sample}

\item{parameter}{parameter of Exponential distribution}
}
\value{
Watson.exp gives Watson statistic of a uniform sample
}
\description{
Compute Watson statistic U^2 for an iid sample, x, to test for the Exponential distribution with parameters unknown.
Estimate parameters by ML using "estimate.exp" by default.
}
\examples{
x= rexp(1000,2)
estimate.exp(x)
Watson.exp(x)
}
