% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.gamma.covmat.R
\name{Watson.gamma.covmat}
\alias{Watson.gamma.covmat}
\title{Watson.gamma.covmat}
\usage{
Watson.gamma.covmat(n, shape)
}
\arguments{
\item{n}{number of eigenvalues}

\item{shape}{shape parameter for Gamma distribution}
}
\value{

}
\description{
Watson.gamma.covmat
}
