% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.gamma.R
\name{Watson.gamma}
\alias{Watson.gamma}
\title{EDF statistics U^2 for Gamma Distribution}
\usage{
Watson.gamma(x, parameter = estimate.gamma(x))
}
\arguments{
\item{x}{random sample}

\item{parameter}{parameter of Gamma distribution}
}
\value{
Watson.gamma gives Watson statistic of a uniform sample
}
\description{
Compute Watson statistic U^2 for an iid sample, x, to test for the Gamma distribution with parameters unknown.
Estimate parameters by ML using "estimate.gamma" by default.
}
\examples{
x= rgamma(1000,1)
estimate.gamma(x)
Watson.gamma(x)
}
