% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.laplace.covmat.R
\name{Watson.laplace.covmat}
\alias{Watson.laplace.covmat}
\title{Watson.laplace.covmat}
\usage{
Watson.laplace.covmat(n)
}
\arguments{
\item{n}{number of eigenvalues}
}
\value{

}
\description{
Watson.laplace.covmat
}
