% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.logistic.covmat.R
\name{Watson.logistic.covmat}
\alias{Watson.logistic.covmat}
\title{Watson.logistic.covmat}
\usage{
Watson.logistic.covmat(n)
}
\arguments{
\item{n}{number of eigenvalues}
}
\value{

}
\description{
Watson.logistic.covmat
}
