% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.normal.covmat.R
\name{Watson.normal.covmat}
\alias{Watson.normal.covmat}
\title{Watson.normal.covmat}
\usage{
Watson.normal.covmat(n)
}
\arguments{
\item{n}{number of eigenvalues}
}
\value{

}
\description{
Watson.normal.covmat
}
