% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.normal.eigen.R
\name{Watson.normal.eigen}
\alias{Watson.normal.eigen}
\title{Watson.normal.eigen}
\usage{
Watson.normal.eigen(n)
}
\arguments{
\item{n}{number of eigenvalues}
}
\value{

}
\description{
Watson.normal.eigen
}
