% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.normal.pvalue.R
\name{Watson.normal.pvalue}
\alias{Watson.normal.pvalue}
\title{P-value of EDF statistics U^2 for Normal Distribution}
\usage{
Watson.normal.pvalue(a, neig = 100, verbose = F)
}
\arguments{
\item{a}{A^2 for Normal Distribution}

\item{neig}{number of eigenvalues}

\item{verbose}{logical; if TRUE, print warning messages}
}
\value{
AD.normal.pvalue gives p-value of the Anderson-Darling statistic of a uniform sample.
}
\description{
Compute p-value of the given Anderson-Darling statistic A^2
}
\examples{
x = rnorm(1000)
asq = AD.normal(x)
AD.normal.pvalue(asq)
}
