% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.regression.pvalue.R
\name{Watson.regression.pvalue}
\alias{Watson.regression.pvalue}
\alias{Watson.normal.regression.pvalue}
\alias{Watson.gamma.regression.pvalue}
\alias{Watson.logistic.regression.pvalue}
\alias{Watson.laplace.regression.pvalue}
\alias{Watson.weibull.regression.pvalue}
\alias{Watson.extremevalue.regression.pvalue}
\alias{Watson.exp.regression.pvalue}
\title{P-value of Watson statistic for regression}
\usage{
Watson.normal.regression.pvalue(w, x, neig = max(n, 400), verbose = FALSE)

Watson.gamma.regression.pvalue(
  w,
  x,
  theta,
  link = "log",
  neig = max(n, 400),
  verbose = FALSE
)

Watson.logistic.regression.pvalue(w, x, neig = max(n, 400), verbose = FALSE)

Watson.laplace.regression.pvalue(w, xneig = max(400, n), verbose = FALSE)

Watson.weibull.regression.pvalue(w, x, neig = max(n, 400), verbose = FALSE)

Watson.extremevalue.regression.pvalue(
  w,
  x,
  neig = max(n, 400),
  verbose = FALSE
)

Watson.exp.regression.pvalue(
  w,
  x,
  theta,
  link = "log",
  neig = max(n, 400),
  verbose = FALSE
)
}
\arguments{
\item{w}{Watson statistic \eqn{U^2} with a given distribution.}

\item{x}{explanatory variables}

\item{verbose}{}
}
\value{

}
\description{

}
\examples{

}
\seealso{

}
