% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.pvalue.R
\name{Watson.uniform.pvalue}
\alias{Watson.uniform.pvalue}
\alias{Watson.normal.pvalue}
\alias{Watson.gamma.pvalue}
\alias{Watson.logistic.pvalue}
\alias{Watson.laplace.pvalue}
\alias{Watson.weibull.pvalue}
\alias{Watson.exp.pvalue}
\title{P-value of Watson statistic}
\usage{
Watson.uniform.pvalue(u, neig = 100, verbose = FALSE)

Watson.normal.pvalue(u, neig = 100, verbose = FALSE)

Watson.gamma.pvalue(u, shape, neig = 100, verbose = FALSE)

Watson.logistic.pvalue(u, neig = 100, verbose = FALSE)

Watson.laplace.pvalue(u, neig = 100, verbose = FALSE)

Watson.weibull.pvalue(u, neig = 100, verbose = FALSE)

Watson.exp.pvalue(u, neig = 100, verbose = FALSE)
}
\arguments{
\item{u}{Watson statistic \eqn{U^2} with a given distribution.}

\item{neig}{Number of eigenvalues used for \code{imhof()}.}

\item{verbose}{Logical; if `TRUE`, print warning messages.}

\item{shape}{The shape parameter of Gamma distribution.}
}
\value{
P-value of the given Watson statistic.
}
\description{
Compute the P-value of the given Watson statistic \eqn{U^2}
using \code{\link[CompQuadForm]{imhof}} function in \code{CompQuadForm}.
}
\examples{
x0=runif(n=100,min=-1,max=1)
usq0 = Watson.uniform(x0)
Watson.uniform.pvalue(usq0)

x1=rnorm(n=100,mean=0,sd=1)
usq1 = Watson.normal(x1)
Watson.normal.pvalue(usq1)

x2=rgamma(n=100,shape=1,scale=1)
usq2 = Watson.gamma(x2)
Watson.gamma.pvalue(usq2,1)

x3=rlogis(n=100,location=0,scale=1)
usq3 = Watson.logistic(x3)
Watson.logistic.pvalue(usq3)

x4= rmutil::rlaplace(n=100,m=0,s=1)
usq4 = Watson.laplace(x4)
Watson.laplace.pvalue(usq4)

x5=rweibull(n=100,shape=1,scale=1)
usq5 = Watson.weibull(x5)
Watson.weibull.pvalue(usq5)

x6=rexp(n=100,rate=1/2)
usq6 = Watson.exp(x6)
Watson.exp.pvalue(usq6)
}
