% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.weibull.covmat.R
\name{Watson.weibull.covmat}
\alias{Watson.weibull.covmat}
\title{Watson.weibull.covmat}
\usage{
Watson.weibull.covmat(n)
}
\arguments{
\item{n}{number of eigenvalues}
}
\value{

}
\description{
Watson.weibull.covmat
}
