% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.pvalue.R
\name{AD.normal.pvalue}
\alias{AD.normal.pvalue}
\alias{AD.gamma.pvalue}
\alias{AD.logistic.pvalue}
\alias{AD.laplace.pvalue}
\alias{AD.weibull.pvalue}
\alias{AD.exp.pvalue}
\title{P-value of the EDF Statistic A^2 for a Given Distribution}
\usage{
AD.normal.pvalue(a, neig = 100, verbose = FALSE)

AD.gamma.pvalue(a, shape, neig = 100, verbose = FALSE)

AD.logistic.pvalue(a, neig = 100, verbose = FALSE)

AD.laplace.pvalue(a, neig = 100, verbose = FALSE)

AD.weibull.pvalue(a, neig = 100, verbose = FALSE)

AD.exp.pvalue(a, neig = 100, verbose = FALSE)
}
\arguments{
\item{a}{Anderson-Darling statistic W^2 for a given distribution.}

\item{neig}{Number of eigenvalues used for \code{imhof()}.}

\item{verbose}{Logical; if TRUE, print warning messages.}

\item{shape}{The shape parameter of Gamma distribution.}
}
\value{
P-value of the Anderson-Darling statistic of a uniform sample.
}
\description{
Compute p-value of the given Anderson-Darling statistic A^2.
}
\details{

}
\examples{
x1=rnorm(n=100,mean=0,sd=1)
asq1 = AD.normal(x1)
AD.normal.pvalue(asq1)

x2=rgamma(n=100,shape=1,scale=1)
asq2 = AD.gamma(x)
AD.gamma.pvalue(asq2,1)

x3=rlogis(n=100,location=0,scale=1)
asq3 = AD.logistic(x)
AD.logistic.pvalue(asq3)

x4= rmutil::rlaplace(n=100,m=0,s=1)
asq4 = AD.laplace(x)
AD.laplace.pvalue(asq4)

x5=rweibull(n=100,shape=1,scale=1)
asq5 = AD.weibull(x)
AD.weibull.pvalue(asq5)

x6=rexp(n=100,rate=1/2)
asq6 = AD.exp(x)
AD.exp.pvalue(asq6)
}
