% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD.regression.pvalue.R
\name{AD.regression.pvalue}
\alias{AD.regression.pvalue}
\alias{AD.normal.regression.pvalue}
\alias{AD.gamma.regression.pvalue}
\alias{AD.logistic.regression.pvalue}
\alias{AD.laplace.regression.pvalue}
\alias{AD.weibull.regression.pvalue}
\alias{AD.extremevalue.regression.pvalue}
\alias{AD.exp.regression.pvalue}
\title{P-value of Anderson-Darling statistic for regression}
\usage{
AD.normal.regression.pvalue(a, x, neig = max(n, 400), verbose = FALSE)

AD.gamma.regression.pvalue(
  a,
  x,
  theta,
  link = "log",
  neig = max(n, 400),
  verbose = FALSE
)

AD.logistic.regression.pvalue(a, x, neig = max(n, 400), verbose = FALSE)

AD.laplace.regression.pvalue(a, x, neig = max(400, n), verbose = FALSE)

AD.weibull.regression.pvalue(a, x, neig = max(n, 400), verbose = FALSE)

AD.extremevalue.regression.pvalue(a, x, neig = max(n, 400), verbose = FALSE)

AD.exp.regression.pvalue(
  a,
  x,
  theta,
  link = "log",
  neig = max(n, 400),
  verbose = FALSE
)
}
\arguments{
\item{x}{explanatory variables}

\item{verbose}{}

\item{w}{Anderson-Darling statistic \eqn{A^2} with a given distribution.}
}
\value{

}
\description{

}
\examples{

}
\seealso{

}
