% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.normal.regression.R
\name{cdf.normal.regression}
\alias{cdf.normal.regression}
\title{Probability Integral Transforms for Normal Regression}
\usage{
cdf.normal.regression(y, x, theta.hat)
}
\arguments{
\item{y}{response variable}

\item{x}{explanatory variables}

\item{theta.hat}{estimates of coefficients}
}
\value{
cdf.normal.regression gives probability integral transforms for y in a ordinary linear regression model
}
\description{
Computes the probability integral transforms of the responses y
in a normal linear regression model
}
