% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.pvalue.R
\name{CvM.normal.pvalue}
\alias{CvM.normal.pvalue}
\alias{CvM.gamma.pvalue}
\alias{CvM.logistic.pvalue}
\alias{CvM.laplace.pvalue}
\alias{CvM.weibull.pvalue}
\alias{CvM.exp.pvalue}
\title{P-value of the EDF Statistic W^2 for a Given Distribution}
\usage{
CvM.normal.pvalue(w, neig = 100, verbose = F)

CvM.gamma.pvalue(w, shape, neig = 100, verbose = FALSE)

CvM.logistic.pvalue(w, neig = 100, verbose = F)

CvM.laplace.pvalue(w, neig = 100, verbose = F)

CvM.weibull.pvalue(w, neig = 100, verbose = FALSE)

CvM.exp.pvalue(w, neig = 100, verbose = FALSE)
}
\arguments{
\item{w}{Cramer-von Mises statistic W^2 for a given distribution.}

\item{neig}{Number of eigenvalues used for \code{imhof()}.}

\item{verbose}{Logical; if TRUE, print warning messages.}

\item{shape}{The shape parameter of Gamma distribution.}
}
\value{
P-value of the Cramer-von Mises statistic of a uniform sample.
}
\description{
Compute p-value of the given Cramer-von Mises statistic W^2.
}
\details{

}
\examples{
x1=rnorm(n=100,mean=0,sd=1)
wsq1 = CvM.normal(x1)
CvM.normal.pvalue(wsq1)

x2=rgamma(n=100,shape=1,scale=1)
wsq2 = CvM.gamma(x)
CvM.gamma.pvalue(wsq2,1)

x3=rlogis(n=100,location=0,scale=1)
wsq3 = CvM.logistic(x)
CvM.logistic.pvalue(wsq)

x4= rmutil::rlaplace(n=100,m=0,s=1)
wsq4 = CvM.laplace(x)
CvM.laplace.pvalue(wsq4)

x5=rweibull(n=100,shape=1,scale=1)
wsq5 = CvM.weibull(x)
CvM.weibull.pvalue(wsq5)

x6=rexp(n=100,rate=1/2)
wsq6 = CvM.exp(x)
CvM.exp.pvalue(wsq6)
}
