% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.regression.statistic.R
\name{CvM.regression}
\alias{CvM.regression}
\alias{CvM.weibull.regression}
\title{Cramer-von Mises statistic for regression}
\usage{
CvM.weibull.regression(y, x, parameter = estimate.weibull.regression(y, x))
}
\arguments{
\item{y}{response variable}

\item{x}{explanatory variables}

\item{parameter}{estimates of regression parameters}
}
\value{

}
\description{

}
\examples{

}
\seealso{

}
