% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvM.regression.statistic.R
\name{CvM.regression}
\alias{CvM.regression}
\alias{CvM.normal.regression}
\alias{CvM.gamma.regression}
\alias{CvM.logistic.regression}
\alias{CvM.laplace.regression}
\alias{CvM.weibull.regression}
\alias{CvM.extremevalue.regression}
\alias{CvM.exp.regression}
\title{Cramer-von Mises statistic for regression}
\usage{
CvM.normal.regression(
  x,
  y,
  fit.intercept = TRUE,
  parameter = estimate.normal.regression(x, y, fit.intercept = fit.intercept)
)

CvM.gamma.regression(
  y,
  x,
  link = "log",
  parameter = estimate.gamma.regression(fit, x, y, link = link)
)

CvM.logistic.regression(y, x, parameter = estimate.logistic.regression(y, x))

CvM.laplace.regression(y, x, parameter = estimate.laplace.regression(y, x))

CvM.weibull.regression(y, x, parameter = estimate.weibull.regression(y, x))

CvM.extremevalue.regression(
  y,
  x,
  parameter = estimate.extremevalue.regression(y, x)
)

CvM.exp.regression(y, x, parameter = estimate.exp.regression(y, x))
}
\arguments{
\item{x}{explanatory variables}

\item{y}{response variable}

\item{parameter}{estimates of regression parameters}
}
\value{

}
\description{

}
\examples{
n = 500
p = 3
beta = c(1,2,3)
x = rnorm(n*(p-1))
x = c(rep(1,n),x)
x = matrix(x,n,p)

# OLS regression
mean = x\%*\%(beta)
#apply the link to get the mean
#generate data with that mean,



sd = 2
y =rnorm(n,mean=mean,sd=sd)
estimate.normal.regression(x=x,y=y)
CvM.normal.regression(x=x,y=y)


}
\seealso{

}
