% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDFtest-package.R
\docType{package}
\name{EDFtest-package}
\alias{EDFtest-package}
\title{EDFtest: Goodness-of-fit Tests based on Empirical Distirbution Function}
\description{
\code{EDFtest} provides software for the calculation of goodness-of-fit test statistics
and their P-values. The three statistics computed are the Empirical Distribution
function statistics called Cramér-von Mises, Anderson-Darling, and Watson statistic.
}
\references{
\url{}.

\url{}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LiYao-sfu/EDFtest}
  \item Report bugs at \url{https://github.com/LiYao-sfu/EDFtest/issues}
}
}
\author{
\strong{Maintainer}: Li Yao \email{yaoliy@sfu.ca}

Authors:
\itemize{
  \item Richard Lockhart
  \item Li Yao
}
}
\keyword{package}
