% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.regression.R
\name{estimate.regression}
\alias{estimate.regression}
\alias{estimate.normal.regression}
\alias{estimate.gamma.regression}
\alias{estimate.exp.regression}
\alias{estimate.laplace.regression}
\alias{estimate.weibull.regression}
\alias{estimate.extremevalue.regression}
\title{Estimate Regression Model Parameters}
\usage{
estimate.normal.regression(x, y, fit.intercept = TRUE)

estimate.gamma.regression(fit, x, y, link = "log")

estimate.exp.regression(fit, x, y, link = "log")

estimate.laplace.regression(x, y, fit.intercept = TRUE)

estimate.weibull.regression(y, x, detail = FALSE)

estimate.extremevalue.regression(y, x, detail = FALSE)
}
\arguments{
\item{x}{A design matrix which is not expected to have a column of 1s.}

\item{y}{A univariate response.}

\item{fit.intercept}{Logical; if TRUE, an intercept term is added by lm.}
}
\value{
Estimated sigma and coefficients of a linear regression.
}
\description{
\code{estimate.normal.regression} fits a standard linear model by OLS (ordinary least square);
\code{estimate.laplace.regression} fits a standard linear model by LAD (least absolute deviations);
\code{estimate.gamma.regression} estimate the parameters in a Gamma regression model
which fits log(E(y))= x beta
}
\examples{
# OLS regression ------------------------------------
n = 500
p = 3
beta = c(1,2,3)
x = rnorm(n*(p-1))
x = c(rep(1,n),x)
x = matrix(x,n,p)
mean = x\%*\%(beta)
sd = 2
y =rnorm(n,mean=mean,sd=sd)
estimate.normal.regression(x,y)

# LAD regression ------------------------------------
y =L1pack::rlaplace(n=n, location=mean, scale=sd)
estimate.laplace.regression(x,y,FALSE)

# Gamma regression ----------------------------------
alpha = 3
scale=exp(x \%*\% beta) / alpha
y =rgamma(n=n,shape=alpha,scale=scale)
estimate.gamma.regression(x,y)
}
\seealso{

}
