% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R, R/gof.normal.R
\name{gof.normal}
\alias{gof.normal}
\alias{gof.normal.bootstrap}
\alias{gof.gamma}
\alias{gof.gamma.bootstrap}
\alias{gof.logistic}
\alias{gof.logistic.bootstrap}
\title{EDF Goodness-of-Fit Tests for a Given Distribution}
\usage{
gof.normal(x, print = TRUE, verbose = FALSE)

gof.normal.bootstrap(x, M = 10000)

gof.gamma(x, print = TRUE, verbose = FALSE)

gof.gamma.bootstrap(x, M = 10000)

gof.logistic(x, print = T, verbose = F)

gof.logistic.bootstrap(x, M = 10000)

gof.normal(x, print = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{random sample}

\item{print}{Logical, if TRUE print statistics and P-values}

\item{verbose}{Logical, if TRUE print every step}

\item{M}{Number of bootstrap, 10000 by default.}
}
\value{
Cramer-von Mises, Anderson-Darling and Watson's statistics and their P-values.

gof.normal computes Cramer-von Mises and Anderson-Darling statistics and their P-values.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate Normal
parameters, compute probability integral transforms, and computes Cramer-von Mises,
Anderson-Darling and Watson's statistics and their P-values using either `imhof`
method or bootstrap.

This function takes in an i.i.d. random sample, use MLE to estimate Normal
parameters, compute probability integral transforms, and computes Cramer-von Mises
and Anderson-Darling statistics and their P-values
}
\details{

}
\examples{
x1=rnorm(n=1000,mean=0,sd=1)
gof.normal(x1)
gof.normal.bootstrap(x1,M=1000)

x2=rgamma(n=1000,shape=1,scale=1)
gof.gamma(x)
gof.normal.bootstrap(x,M=1000)

x3=rlogis(n=1000,location=0,scale=1)
usq3 = Watson.logistic(x)
Watson.logistic.pvalue(usq3)

x4= rmutil::rlaplace(n=1000,m=0,s=1)
gof.laplace(x4)
gof.laplace.bootstrap(x4,M=1000)
x = rnorm(100,1)
gof.normal(x)
gof.normal(x,print=TRUE,verbose=TRUE)
}
