% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.sandwich.R
\name{gof.sandwich}
\alias{gof.sandwich}
\title{EDF Goodness-of-Fit tests for General Distributions using Sandwich Estimation of Covariance Function}
\usage{
gof.sandwich(y, x = NULL, Fdist, thetahat, Score, m = max(n, 100), ...)
}
\arguments{
\item{y}{data -- should be a numerical vector: sample or response of regression problem}

\item{x}{matrix of covariates}

\item{Fdist}{user supplied function to compute probability integral transform of y}

\item{thetahat}{parameter estimates by mle}

\item{Score}{user supplied function to compute3 components of the score function an n by p matrix with entries
partial log f(y_i,\theta)/ partial theta_j}

\item{m}{Eigenvalues are extracted for an m by m grid of the covariance ftn}

\item{...}{other inputs passed to Fdist and Score when needed.}
}
\value{

}
\description{
This function tests the hypothesis that data y come from
distribution Fdist with unknown parameter values theta
}
\details{
Estimates of theta must be provided in thetahat.

It uses a large sample approximation to the limit distribution
based on the use of the score function components
to estimate the Fisher information and the limiting covariance
function of the empirical process.

The estimates thetahat should be roots of the likelihood equations.
}
\examples{
sample = rnorm(n=100,mean=0,sd=1)
mle = estimate.normal(sample)
cdf.normal.user = function(x,theta){
 pnorm(x,mean=theta[1],sd=theta[2])
}
score.normal.user = function(x,theta){
 sig=theta[2]
 mu=theta[1]
 s.mean= (x-mu)/sig
 s.sd= s.mean^2/sig-length(x)/sig
 cbind(s.mean/sig,s.sd)
}
output = gof.sandwich(y=sample,Fdist=cdf.normal.user,thetahat=mle,Score=score.normal.user,m=100)
output
#do a mc study;plot two pvalues for two methods. each pvalue use AD.uniform.pvalue=.05-0.95 ;

}
