% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.statistics.only.R
\name{gof.statistics.only}
\alias{gof.statistics.only}
\title{Goodness-of-Fit tests for statistics only}
\usage{
gof.statistics.only(pit, AD = TRUE, CvM = TRUE, Watson = TRUE)
}
\arguments{
\item{pit}{A vector of probability integral transforms.}

\item{AD}{Logical; if TRUE print Anderson-Darling statistic.}

\item{CvM}{Logical; if TRUE print Cramér-von Mises statistic.}

\item{Watson}{Logical; if TRUE print Watson statistic.}
}
\value{
`gof.statistics.only` gives Anderson-Darling, Cramér-von Mises or Watson statistics
of a given probability integral transforms
}
\description{
This function takes in a vector of probability integral transforms
and computes whichever of the three statistics is asked for.
By default {A^2},{W^2}, and {U^2} are computed
}
\examples{
x= runif(100)
gof.statistics.only(x)
}
