% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{gof.uniform}
\alias{gof.uniform}
\alias{gof.normal}
\alias{gof.gamma}
\alias{gof.logistic}
\alias{gof.laplace}
\alias{gof.weibull}
\alias{gof.exp}
\title{Generic GOF tests based on EDF}
\usage{
gof.uniform(x, print = TRUE, verbose = FALSE)

gof.normal(x, print = TRUE, verbose = FALSE)

gof.gamma(x, print = TRUE, verbose = FALSE)

gof.logistic(x, print = TRUE, verbose = FALSE)

gof.laplace(x, print = TRUE, verbose = FALSE)

gof.weibull(x, print = TRUE, verbose = FALSE)

gof.exp(x, print = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{A random sample.}

\item{print}{Logical; if TRUE print both statistics and P-values.}

\item{verbose}{verbose Logical; if TRUE, print warning messages.}
}
\value{
Cramér-von Mises, Anderson-Darling and Watson statistics and their P-values.
}
\description{
This function takes in an i.i.d. random sample, use MLE to estimate parameters of
the assumed distribution, compute probability integral transforms, and computes Cramér-von Mises,
Anderson-Darling and Watson statistics and their P-values using \code{imhof()}
function in \code{CompQuadForm}.
}
\examples{
x0=runif(n=100,min=-1,max=1)
gof.uniform(x0)

x1=rnorm(n=100,mean=0,sd=1)
gof.normal(x1)

x2=rgamma(n=100,shape=1,scale=1)
gof.gamma(x2)

x3=rlogis(n=100,location=0,scale=1)
gof.logistic(x3)

x4= rmutil::rlaplace(n=100,m=0,s=1)
gof.laplace(x4)

x5=rweibull(n=100,shape=1,scale=1)
gof.weibull(x5)

x6=rexp(n=100,rate=1/2)
gof.exp(x6)
}
