% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.pvalue.R
\name{Watson.normal.pvalue}
\alias{Watson.normal.pvalue}
\title{P-value of EDF statistics U^2 for Normal Distribution}
\usage{
Watson.normal.pvalue(u, neig = 100, verbose = FALSE)
}
\arguments{
\item{u}{Watson's statistic U^2 for a given distribution.}

\item{neig}{Number of eigenvalues used for \code{imhof()}.}

\item{verbose}{Logical; if TRUE, print warning messages.}

\item{shape}{The shape parameter of Gamma distribution.}
}
\value{
P-value of the Watson's statistic of a uniform sample.
}
\description{
Compute p-value of the given Anderson-Darling statistic A^2
}
\details{

}
\examples{
x1=rnorm(n=100,mean=0,sd=1)
usq1 = Watson.normal(x1)
Watson.normal.pvalue(usq1)

x2=rgamma(n=100,shape=1,scale=1)
usq2 = Watson.gamma(x)
Watson.gamma.pvalue(usq2,1)

x3=rlogis(n=100,location=0,scale=1)
usq3 = Watson.logistic(x)
Watson.logistic.pvalue(usq3)

x4= rmutil::rlaplace(n=100,m=0,s=1)
usq4 = Watson.laplace(x)
Watson.laplace.pvalue(usq4)

x5=rweibull(n=100,shape=1,scale=1)
usq5 = Watson.weibull(x)
Watson.weibull.pvalue(usq5)

x6=rexp(n=100,rate=1/2)
usq6 = Watson.exp(x)
Watson.exp.pvalue(usq6)
}
