% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Watson.regression.R
\name{Watson.regression}
\alias{Watson.regression}
\alias{Watson.normal.regression}
\title{Watson statistic for regression}
\usage{
Watson.normal.regression(y, x, parameter = estimate.normal.regression(y, x))
}
\arguments{
\item{y}{response variable}

\item{x}{explanatory variables}

\item{parameter}{estimates of regression parameters}
}
\value{

}
\description{

}
\examples{

}
\seealso{

}
